Beginning a custom project starter repository
I am making this repository to build up a blank 'starter file' for projects.

My first include here is Grunt. At the command line, I ran:

`npm install -g grunt-cli`

I am following the directions found here:
https://gruntjs.com/getting-started

I noticed on other repos I was using that once I installed `grunt`,
Github pages would not build because of things in the node_modules
folder throwing errors. We'll see if this is the case.

My goal is a folder that lets me see my files using a local
webserver and be able to push changes to a Github Pages-backed site
when I am ready. If `grunt` is not the answer, I might use
python SimpleHTTP for the localhost webserver.

Next step: add package.json and a Gruntfile via the instructions
found at the site above.