Adding package.json and Gruntfile to create Grunt services
Adding the following files:

`gruntfile.js`
`package.json`

according to the instructions found here:

https://gruntjs.com/getting-started

The goal here is to get `grunt serve` to work.

If I remember correctly, I will end up having to add a few
things to the package.json to get it to work. May need to refer to
other gruntfile/package.json. However, want to keep this as clean
and simple as possible to maximize chances of a clean Github Pages build.
If we drag in too many node_modules, the build will probably fail.

Next step: `npm install` and add an index.html file to test the local
and remote page builds.